requirements-dev.txt
# Development and testing dependencies for AgentNet
# Install with: pip install -r requirements-dev.txt

# Core dependencies (from registry service - FastAPI based)
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.0
python-multipart==0.0.6
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2
email-validator>=2.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Linting & Formatting
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
